Extract text utils that is used by `babel-parser.js`
