+++ src/common/util.js
@@ -5,6 +5,16 @@ const getLast = require("../utils/get-last.js");
 const { getSupportInfo } = require("../main/support.js");
 const isNonEmptyArray = require("../utils/is-non-empty-array.js");
 const getStringWidth = require("../utils/get-string-width.js");
+const {
+  skipWhitespace,
+  skipSpaces,
+  skipToLineEnd,
+  skipEverythingButNewLine,
+} = require("../utils/text/skip.js");
+const skipInlineComment = require("../utils/text/skip-inline-comment.js");
+const skipTrailingComment = require("../utils/text/skip-trailing-comment.js");
+const skipNewline = require("../utils/text/skip-newline.js");
+const getNextNonSpaceNonCommentCharacterIndexWithStartIndex = require("../utils/text/get-next-non-space-non-comment-character-index-with-start-index.js");
 
 const getPenultimate = (arr) => arr[arr.length - 2];
 
@@ -51,110 +61,6 @@ function skip(chars) {
     }
     return false;
   };
-}
-
-/**
- * @type {(text: string, index: number | false, opts?: SkipOptions) => number | false}
- */
-const skipWhitespace = skip(/\s/);
-/**
- * @type {(text: string, index: number | false, opts?: SkipOptions) => number | false}
- */
-const skipSpaces = skip(" \t");
-/**
- * @type {(text: string, index: number | false, opts?: SkipOptions) => number | false}
- */
-const skipToLineEnd = skip(",; \t");
-/**
- * @type {(text: string, index: number | false, opts?: SkipOptions) => number | false}
- */
-const skipEverythingButNewLine = skip(/[^\n\r]/);
-
-/**
- * @param {string} text
- * @param {number | false} index
- * @returns {number | false}
- */
-function skipInlineComment(text, index) {
-  /* istanbul ignore next */
-  if (index === false) {
-    return false;
-  }
-
-  if (text.charAt(index) === "/" && text.charAt(index + 1) === "*") {
-    for (let i = index + 2; i < text.length; ++i) {
-      if (text.charAt(i) === "*" && text.charAt(i + 1) === "/") {
-        return i + 2;
-      }
-    }
-  }
-  return index;
-}
-
-/**
- * @param {string} text
- * @param {number | false} index
- * @returns {number | false}
- */
-function skipTrailingComment(text, index) {
-  /* istanbul ignore next */
-  if (index === false) {
-    return false;
-  }
-
-  if (text.charAt(index) === "/" && text.charAt(index + 1) === "/") {
-    return skipEverythingButNewLine(text, index);
-  }
-  return index;
-}
-
-// This one doesn't use the above helper function because it wants to
-// test \r\n in order and `skip` doesn't support ordering and we only
-// want to skip one newline. It's simple to implement.
-/**
- * @param {string} text
- * @param {number | false} index
- * @param {SkipOptions=} opts
- * @returns {number | false}
- */
-function skipNewline(text, index, opts) {
-  const backwards = opts && opts.backwards;
-  if (index === false) {
-    return false;
-  }
-
-  const atIndex = text.charAt(index);
-  if (backwards) {
-    // We already replace `\r\n` with `\n` before parsing
-    /* istanbul ignore next */
-    if (text.charAt(index - 1) === "\r" && atIndex === "\n") {
-      return index - 2;
-    }
-    if (
-      atIndex === "\n" ||
-      atIndex === "\r" ||
-      atIndex === "\u2028" ||
-      atIndex === "\u2029"
-    ) {
-      return index - 1;
-    }
-  } else {
-    // We already replace `\r\n` with `\n` before parsing
-    /* istanbul ignore next */
-    if (atIndex === "\r" && text.charAt(index + 1) === "\n") {
-      return index + 2;
-    }
-    if (
-      atIndex === "\n" ||
-      atIndex === "\r" ||
-      atIndex === "\u2028" ||
-      atIndex === "\u2029"
-    ) {
-      return index + 1;
-    }
-  }
-
-  return index;
 }
 
 /**
@@ -232,26 +138,6 @@ function isNextLineEmptyAfterIndex(text, index) {
  */
 function isNextLineEmpty(text, node, locEnd) {
   return isNextLineEmptyAfterIndex(text, locEnd(node));
-}
-
-/**
- * @param {string} text
- * @param {number} idx
- * @returns {number | false}
- */
-function getNextNonSpaceNonCommentCharacterIndexWithStartIndex(text, idx) {
-  /** @type {number | false} */
-  let oldIdx = null;
-  /** @type {number | false} */
-  let nextIdx = idx;
-  while (nextIdx !== oldIdx) {
-    oldIdx = nextIdx;
-    nextIdx = skipSpaces(text, nextIdx);
-    nextIdx = skipInlineComment(text, nextIdx);
-    nextIdx = skipTrailingComment(text, nextIdx);
-    nextIdx = skipNewline(text, nextIdx);
-  }
-  return nextIdx;
 }
 
 /**
