+++ src/utils/text/skip-newline.js
@@ -1,0 +1,54 @@
+"use strict";
+
+/** @typedef {import("./skip").SkipOptions} SkipOptions */
+
+// This one doesn't use the above helper function because it wants to
+// test \r\n in order and `skip` doesn't support ordering and we only
+// want to skip one newline. It's simple to implement.
+/**
+ * @param {string} text
+ * @param {number | false} index
+ * @param {SkipOptions=} opts
+ * @returns {number | false}
+ */
+function skipNewline(text, index, opts) {
+  const backwards = opts && opts.backwards;
+  if (index === false) {
+    return false;
+  }
+
+  const atIndex = text.charAt(index);
+  if (backwards) {
+    // We already replace `\r\n` with `\n` before parsing
+    /* istanbul ignore next */
+    if (text.charAt(index - 1) === "\r" && atIndex === "\n") {
+      return index - 2;
+    }
+    if (
+      atIndex === "\n" ||
+      atIndex === "\r" ||
+      atIndex === "\u2028" ||
+      atIndex === "\u2029"
+    ) {
+      return index - 1;
+    }
+  } else {
+    // We already replace `\r\n` with `\n` before parsing
+    /* istanbul ignore next */
+    if (atIndex === "\r" && text.charAt(index + 1) === "\n") {
+      return index + 2;
+    }
+    if (
+      atIndex === "\n" ||
+      atIndex === "\r" ||
+      atIndex === "\u2028" ||
+      atIndex === "\u2029"
+    ) {
+      return index + 1;
+    }
+  }
+
+  return index;
+}
+
+module.exports = skipNewline;
